api-ref: Replace non UUID string with UUID
In the "List Migrations" (GET /os-migrations) API,
there are the following response body examples.

* Example List Migrations: JSON response
* Example List Migrations (v2.59)
* Example List Migrations With Paging (v2.59)

In the second and third examples, the 'instance_uuid' are UUID strings.
But in the first one, the 'instance_uuid' is not UUID.
It should be a UUID string.
So replace non UUID string with a UUID string in the example.

Change-Id: I6c29c48f53838beb7da3ec0f849380e3706df422