Ignore hw_vif_type for direct, direct-physical vNIC types
As part of the fix for bug #1789074, we stopped setting the VIF model to
virtio for a vif_type of 'direct' or 'direct-physical'. However, if
users configured the 'hw_vif_type' image metadata property, the VIF
model would be unconditionally set to whatever the user provided. This
would result in bug #1789074 recurring. Given that the model attribute
is unnecessary for these vNIC types, we can resolve this issue by
simply ignoring the image metadata property when configuring such VIFs.

Change-Id: I3bf5587c3a49c807bb14533dbbea8f72dca2f02b
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>
Closes-bug: #1797146