Handle volume API failure in _post_live_migration
Previously, if the call to Cinder in _post_live_migration failed, the
exception went unhandled and prevented us from calling
post_live_migration_at_destination - which is where we set instance
host and task state. This left the system in an inconsistent state,
with the instance actually running on the destination, but
with instance.host still set to the source. This patch simply wraps
the Cinder API calls in a try/except, and logs the exception instead
of blowing up. While "dumb", this has the virtue of being simple and
minimizing potential side effects. A comprehensive refactoring of
when, where and how we set instance host and task state to try to
guarantee consistency is left as a TODO.

Partial-bug: 1628606
Change-Id: Icb0bdaf454935b3713c35339394d260b33520de5
(cherry picked from commit 5513f48dea529fe4e690f50a462300129594210c)