Fix NoneType error in _notify_volume_usage_detach
If the driver.block_stats() method returns None, like the
libvirt driver will if the guest is gone during an evacuate,
we'll get a NoneType error trying to unpack the return value
from the driver. Instead, simply return as if the driver
raised NotImplementedError.

Since handling None is changing the contract on the virt
driver API, the docstring is updated to explain the acceptable
return values of the driver method.

Change-Id: I98a2785c07f7af02ad83650c72d9e1868290ece4
Closes-Bug: #1796981