Explain nested guest support
In the Configuration Guide's section on KVM:

* expand on the implications of selecting a CPU mode and model
  for live migration,
* discuss how to enable nested guests, and the implications and
  limitations of doing so,
* bump the heading level of "Guest agent support".

This patch differs from the version in Stein
(e304ad7f4d3bf0d0cf4e5b4d2de3f2c4be2f2b8e), and the clean cherry-pick
into Rocky (63bf3833ae40d09c9a5e1ea9056c20d041da4fa5) in that it omits
mentioning the option of enabling the vmx flag via
cpu_model_extra_flags. The rationale for that is that vmx is not a
supported option in stable/queens at the time of writing (only pcid,
ssbd, virt-ssbd, amd-ssbd, and amd-no-ssb are), so the only option to
enable nested guest support in Queens (and earlier releases) is to use
cpu_mode=host-model or cpu_mode=host-passthrough, and enable nested
virtualization via the kernel module.

Using cpu_mode=custom is not an option for enabling nested
virtualization in Queens in the absence of vmx as an available option
for cpu_model_extra_flags, regardless of the cpu_model being used,
because (again at the time of writing) none of the CPU models defined
in libvirt do enable the vmx flag automatically.

Closes-Bug: 1791678
Change-Id: I671acd16c7e5eca01b0bd633caf8e58287d0a913