Add regression test for bug 1797580
Microversion 2.56 allows cold migrating to a specified target host. The
compute API sets the requested destination on the request spec with the
specified target host and then conductor sends that request spec to the
scheduler to validate the host. Conductor later persists the changes to
the request spec because it's the resize flow and the flavor could change
(even though in this case it won't since it's a cold migrate). After
confirming the resize, if the server is live migrated it will fail during
scheduling because of the persisted RequestSpec.requested_destination
from the cold migration, and you can't live migrate to the same host on
which the instance is currently running.

This change adds a test to recreate the regression bug.

Change-Id: I588655fdd90917d00ccf5eb0a8df7bccc1ac0e81
Related-Bug: #1797580