Move live_migration.pre.start to the start of the method
This simply moves the pre_live_migration start notification
back to the start of the method before creating new volume
attachments. While in here, the list of bdms are passed to
the notify_about_instance_action method so if
[notifications]/bdms_in_notifications is True we don't have
to pull them out of the database again for each versioned
notification.

Change-Id: I372eddb9e356b02328f937917a856bc631691f53
Partial-Bug: #1763051