Fix best_match() deprecation warning
wsgi call to best_match() generates the following warning:

DeprecationWarning: The behavior of .best_match for the Accept classes
is currently being maintained for backward compatibility,
but the method will be deprecated in the future,
as its behavior is not specified in
(and currently does not conform to) RFC 7231.

See the following URL for more information about deprecation.

https://docs.pylonsproject.org/projects/webob/en/stable/api/webob.html#webob.acceptparse.AcceptValidHeader.best_match

Simlar changes in the wsgi best language code as well,
except new call is to lookup().

Change-Id: Ib93ffffd3f28aa6d53a0a4eed5aeee94900cb073
Closes-Bug: #1798224