Add nova-status upgrade check for consoles
This will check if a deployment is currently using consoles and warns
the operator to set [workarounds]enable_consoleauth = True on their
console proxy host if they are performing a rolling upgrade which is
not yet complete.

Partial-Bug: #1798188

 Conflicts:
	doc/source/cli/nova-status.rst

NOTE(melwitt): The conflict is from the absence of the
**19.0.0 (Stein)** section in Rocky and the absence of the placement
API check to require version 1.30 in Rocky.

Change-Id: Idd6079ce4038d6f19966e98bcc61422b61b3636b
(cherry picked from commit d2535b0261205ed80e1731451496caff19ef3f63)