Make CellDatabases fixture reentrant
The current CellDatabases fixture uses a single lock to protect code
that needs to run with the global database state pointed at a specific
cell. That is a problem if we ever need to recursively target a cell
(which would work in real life but not in tests), as well as when we have
two threads that need to target a cell where other locks are involved
such that they deadlock each other.

This change attempts to convert the fixture to use the existing
writer lock to only protect the code that changes the global database
state, and a reader lock for all threads actually running code that
is targeted. This avoids the deadlock because the schedulable code that
could acquire other locks isn't blocking other threads, except those that
need to change to another cell.

Change-Id: I9cd701dea7356bd3825900e216b163b895299e08