Drop legacy cold migrate allocation compat code
Migration-based allocations, where conductor swaps the
source node allocations from the instance record to the
migration record before calling the scheduler - which
creates new allocations on the selected dest node for
the instance, was new in Queens but contained compatibility
code for (1) old computes where conductor couldn't pass
a migration record to the compute and (2) scheduler drivers
which didn't create allocations, like the CachingScheduler.

Change Ibcb6bf912b3fb69c8631665fef2832906ba338aa makes
"migration" a required parameter for prep_resize() in
compute so (1) is no longer an issue. And change
I1832da2190be5ef2b04953938860a56a43e8cddf removed the
CachingScheduler.

This change guts the compatibility code added throughout
the following changes in Queens:

  I0883c2ba1989c5d5a46e23bcbcda53598707bcbc

  I7b2903c56cb53b48afe2c4dec42ae7623a7f24eb

  I7c3c95d4e0836e1af054d076aad29172574eab2c

This change does overlap a bit with live-migration
flows that deal with migration-based allocations. Those
flows will be cleaned up in a subsequent change.

Cleaning up this compatibility code is going to help
with supporting nested resource allocations for things
like blueprint use-nested-allocation-candidates where
we need to be sure that the source node allocations are
on the migration record rather than "doubled up" on the
instance record.

Change-Id: I0851e2d54a1fdc82fe3291fb7e286e790f121e92