Fix min config value for shutdown_timeout option
The shutdown_timeout config option was added in commit
c07ed15415c0ec3c5862f437f440632eff1e94df without a min
value. The min value was later added in commit
d67ea6e5549086eee1b39946648410f22d0041a9 and set to 1,
which means the option can never be configured to mean
"always immediately shutdown". That is also inconsistent
with the description of the "os_shutdown_timeout" image
property in the glance metadata definition which says the
value can be set to 0 to force an immediate shutdown of
the guest.

This fixes the min value in the config option to be 0 which
is already what happens if we are not performing a clean
shutdown at all.

Change-Id: I399b9031d2aa477194697390e2cd3f78e3ac0f91
Closes-Bug: #1799707