Add ratio online data migration when load compute node
This patch adds an online data migrations for any compute_nodes
with existing ``0.0`` and ``None`` allocation ratio. If it's an
existing record with 0.0 values, we will replace it with the
configure ``xxx_allocation_ratio`` config if it's not None, and
fallback to using the ``initial_xxx_allocation_ratio`` otherwise.

Change-Id: Ic137d837c7b1d29f56a45a01b3b104d6d2c698df
blueprint: initial-allocation-ratios