tox: Stop build *all* docs in 'docs'
The 'docs' target currently builds the documentation trees in 'api-ref',
'api-guide' and 'placement-api-ref', in addition to 'doc'. This
massively increases the amount of time docs take to build both locally
and in the gate. It's not necessary for gate, since separate jobs take
care of the other documents for us [1]. As such, we should stop doing
it. For users that *do* care about this (for whatever reason) a new
'all-docs' target is included.

[1] https://github.com/openstack-infra/project-config/blob/master/zuul.d/projects.yaml#L5578-L5595

Change-Id: I58bd50e869fc00dde5dd388efb686a7196c8db80
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>
(cherry picked from commit f49b81fa2bd692583370bab569f25215073a5bc2)