Add tests for bug #1800511
In Rocky, we started including the MTU for networks when vif_type
is bridge or tap and libvirt >= 3.3.0, however this also meant
that we started specifying the MTU when doing live migrations.

It seems that the guest ABI changes when setting the MTU which
means that live migrations will fail.  This broke the live migration
of any instance that was launched prior to upgrading to Rocky, as
it was not loaded with the ABI having host_mtu specified.

This is a passing (negative) test-case which will include a follow
up patch that contains a fix and correcting tests.

(cherry picked from commit 29ee8011b4e1cf5371b1aa9c6c0f930eb49fe795)

Related-Bug: #1800511
Change-Id: Ia2fe50d727b1f83e808cb9dda3a55f853f048a3e