Add Migration.cross_cell_move and get_by_uuid
This adds a new boolean column, cross_cell_move,
to the migrations table in the cell database and
corresponding field in the Migrations versioned
object. This will be used to track cross-cell
resize operations. Also, a Migration.get_by_uuid()
method is added to the versioned object because
while doing a cross-cell resize, we will need to
be able to lookup a migration record via its uuid
which will be unique across cell databases, unlike
its id value which is specific to each cell database.

Note that we opt to not store the source/dest cell
UUIDs directly in the migrations table because the
migrations records are stored in the cell database and
we prefer to keep the cell information (which is in the
API) separated from the cell database, especially in
case the cell UUID maps to something that does not
exist later (like if the cell is split).

Part of blueprint cross-cell-resize

Change-Id: I63411214c412af93501b27eb39f97bf980b21760