Add Destination.allow_cross_cell_move field
This adds a new boolean field, allow_cross_cell_move, to
the Destination object used within a RequestSpec.

This will be used during a cross-cell resize when asking
the scheduler for a target host whether or not that host
should be restricted to the current cell or if the host
can come from another cell. Furthermore, it will be used
in a scheduler weigher which will prefer candidate target
hosts in the current cell rather than a different cell.

Part of blueprint cross-cell-resize

Change-Id: I6e226321348f5d64d16b0bae26ad31c52debf0bc