Add InstanceAction/Event create() method
During a cross-cell resize, the instance and its
related records are created in the target cell. Since
the resize is initiated in the source cell, instance
actions and associated events will be created in the
source cell and need to be re-created in the target cell
where the resize completes. Not only that, but we want
to also make sure the history of actions performed on
an instance are copied into the target cell database
in case the resize is confirmed there.

This adds a create() method for the InstanceAction
and InstanceActionEvent objects to allow directly
re-creating those resources from the source cell into
the target cell.

The existing action_start() and event_start() methods
will not work for this scenario since those depend on the
current context but we need to clone actions/events created
with an older/different context.

Part of blueprint cross-cell-resize

Change-Id: I43db251f708c18eb35efe1fdbcdada35ae7e792c