Change HostManager to allow scheduling to other cells
As part of the cross-cell resize work, we need to be able
to tell the scheduler that it's OK to select destinations
from cells in which the instance does not currently live.
This happens via the "allow_cross_cell_move" flag in the
RequestSpec.requested_destination field, which is not set
at this time (this is just plumbing which will be used later)
and defaults to False for backward compatibility.

Part of blueprint cross-cell-resize

Change-Id: Ib29a17f7fcc3280871d5c9ba31e1374eff98cf82