quota: remove defaults kwarg in get_project_quotas
Both QuotaDriver.get_project_quotas() and QuotaDriver.get_user_quotas()
had a default kwarg that was never used. The only call sites for this
method are in the limits and quota-sets API handlers and none of the
call sites overrides the defaults kwarg. So, remove it. It was only
cluttering up the interface.

Change-Id: I0b7acbe7bd818ef313aefcc5f8d58277d101ce3f