Update compute API.get() mocks in test_server_metadata
These old unit tests are stubbing out the DB API to
get an instance which is way too low level. This change
updates the stubs to stub out the compute API.get method
instead.

While in here, some old unused stubs are removed and
common fakes code is used.

This is part of a larger series to eventually
drop pre-cellsv2 compatibility code in API.get().

Change-Id: I9542c66b3f6f52a7d29046ba3897ca28fb8e377b