Update compute API.get() stubs for test_*security_groups
These old unit tests were stubbing the DB API to get
instance records which is too low level for what we're
testing in the API. This changes the stubs to be on the
compute API.get() method instead.

While in here, tests are changed to re-use common stub
code and some unnecessary stubs/tests are removed. Notably
the non-running and 'invalid instance' tests are removed
since they don't test anything useful that is not already
tested (power state doesn't matter for associating/disassociating
security groups to an instance and there are already tests
for instances which don't exist).

This is part of a larger series of changes to drop some
pre-cellsv2 compatibility code from compute API.get().

Change-Id: I203dc5f2abc14c7ab7c140f6a1e35c2a433b9568