Update compute API.get() stubs for test_disk_config
These old unit tests were stubbing the DB API to get
an instance which is too low level. This change updates
the stub to be on the compute API.get() method. While
in here several unnecessary stubs are also removed.

This is part of a larger series of changes to drop
pre-cellsv2 compat code in the compute API.get() method.

Change-Id: I623e2ed17013ee8fae22a1527199412ded4542b4