Drop pre-cellsv2 compat in compute API.get()
This removes the nova-osapi_compute service version
compatibility check for cells v2 which was added in
Newton.

A "nova-status upgrade check" was added for that
in Rocky:

  Ie2bc4616439352850cf29a9de7d33a06c8f7c2b8

And it is also causing issues with functional tests
where instances are buried in cell0 and because we
don't use the instance mapping, we fail to find the
instance.

As a result of this, compute API tests that rely on
retrieving a real instance out of the database will
also need to make sure an associated instance mapping
record exists.

Change-Id: I11083aa3c78bd8b6201558561457f3d65007a177
Closes-Bug: #1800472