Separate methods to free claimed and allocated devs
The existing PCI manager exposes a single method to free PCI resources
free_instance(), which frees both claimed and allocated PCI resources
for an instance.

This change proposes to extend the PCI manager API with two methods:
  1. free_instance_claims() : free PCI resources claims for instance.
  2. free_instance_allocations() : free PCI resources allocations for
     instance.

This change refactors free_instance() to use (1) and (2) from above.

This change is required to enable SR-IOV live migration as it is
required to free instance PCI allocations on the source node
in case of a successful migration and free instance PCI claims
on the destination node in case of an unsuccessful migration.

Change-Id: Id961f0fc219f32a2cf0282859f228e87cb36ffeb
Partial-Implements: blueprint libvirt-neutron-sriov-livemigration