Convert exception messages to strings
Exception messages are expected to be strings, but it's not always
the case. Some exceptions are wrapped and passed as new exception
messages, which affects some exception handlers.

Finding all the occurrences may be difficult. For now, it's easier
to just convert the exception messages to strings in the NovaException
class.

Closes-Bug: #1801713

Change-Id: Ic470791aa2b08ca9911bf70cb3cc68652d3647f2