Remove LazyLoad of Scheduler Clients
Things have changed here on Walton's Mountain since LazyLoad was
introduced [1]. It seems to have been created to avoid a circular
import, but as this patch should attest, that's no longer an issue.

Why change this now, besides removing weird and complicated code?
Because a subsequent patch needs to access a *property* of the report
client from the compute manager. As written, LazyLoad only lets you get
at *methods* (as functools.partial). There are other ways to solve that
while preserving the deferred import, but if this works, it's the better
solution.

[1] Ie5732baf9709cd0cb951eae4638910372c79e5f1

Change-Id: I1f97d00fb7633f173370ed6787c9a71ecd8106d5