SIGHUP n-cpu to clear provider tree cache
An earlier change [1] allowed
[compute]resource_provider_association_refresh to be set to zero to
disable the resource tracker's periodic refresh of its local copy of
provider traits and aggregates. To allow for out-of-band changes to
placement (e.g. via the CLI) to be picked up by the resource tracker in
this configuration (or a configuration where the timer is set to a high
value) this change clears the provider tree cache when SIGHUP is sent to
the compute service. The next periodic will repopulate it afresh from
placement.

[1] Iec33e656491848b26686fbf6fb5db4a4c94b9ea8

Change-Id: I65a7ee565ca5b3ec6c33a2fd9e39d461f7d90ed2