Add functional test to delete a server while in VERIFY_RESIZE
This scenario came up while discussing what might be causing
leaked resource allocations in Placement [1].

I fully expected the new test to fail because I couldn't see
from the compute manager where the migration-based allocations
would be cleaned up when deleting the server. It turns out that
when deleting a VERIFY_RESIZE server, the API confirms the resize
which drops the migration-based allocations on the source node
before deleting the server on the target node.

Since this is not obvious, a comment in the compute API
_confirm_resize_on_deleting() method is added.

[1] http://lists.openstack.org/pipermail/openstack-dev/2018-November/136295.html

Change-Id: I4c12502c86c7ac27369d119e0f97768cf41695b5